requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1; python_version<"3.9"
beautifulsoup4==4.10.0
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.1
django-admin-rangefilter==0.8.7
django-extensions==3.2.0
djangorestframework==3.13.1
drf-api-logger==1.1.10
idna==3.3
kombu==5.2.4
numpy==1.23.2
packaging==21.3
phonenumbers==8.12.53
pip-autoremove==0.10.0
pipdeptree==2.2.1
prompt-toolkit==3.0.30
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.2.1
redis==4.3.4
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
